refactor: restructure voice_generator module and enhance GUI integration
Refactored the voice_generator module by consolidating audio processing functions and introducing a new GUI structure using Gradio. The main application logic has been moved to a dedicated `gui.py` file, improving code organization. Updated the `pyproject.toml` to include build system requirements and project scripts. Enhanced error handling and streamlined the import statements for better clarity. Additionally, updated the .gitignore to exclude all __pycache__ directories.
