+++ .gitignore
@@ -2,5 +2,4 @@ tmp/
 .venv/
 .ruff_cache/
 .mypy_cache/
-__pycache__/
-src/voice_generator/__pycache__/
+**/__pycache__/
