+++ .idea/MypyPlugin.xml
@@ -2,6 +2,7 @@
 <project version="4">
   <component name="MypySettings">
     <option name="arguments" value="--follow-imports silent --exclude \.pyi$" />
+    <option name="mypyExecutable" value="D:\Dev\chatacter_backend\VoiceGenerator\.venv\Scripts\mypy.exe" />
     <option name="projectDirectory" value="$PROJECT_DIR$" />
   </component>
 </project>
