+++ .idea/VoiceGenerator.iml
@@ -5,4 +5,10 @@
     <orderEntry type="jdk" jdkName="uv (VoiceGenerator)" jdkType="Python SDK" />
     <orderEntry type="sourceFolder" forTests="false" />
   </component>
+  <component name="PackageRequirementsSettings">
+    <option name="requirementsPath" value="" />
+  </component>
+  <component name="PyDocumentationSettings">
+    <option name="renderExternalDocumentation" value="true" />
+  </component>
 </module>
